split up webpack commons into categories
